requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pdfminer.six==20231228
slowapi==0.1.9